working appointment creation with front end appointment-checking function to conditionally render available time slots and back end validations to check if appointment has valid times

+++ backend/db/models/appointment.js
@@ -4,8 +4,8 @@ module.exports = (sequelize, DataTypes) => {
     "Appointment",
     {
       date: DataTypes.INTEGER,
-      startTime: DataTypes.INTEGER,
-      hours: DataTypes.INTEGER,
+      startTime: DataTypes.NUMERIC,
+      hours: DataTypes.NUMERIC,
       employeeId: DataTypes.INTEGER,
       customerId: DataTypes.INTEGER,
       services: DataTypes.STRING,
