+++ backend/db/models/service.js
@@ -5,8 +5,8 @@ module.exports = (sequelize, DataTypes) => {
     {
       title: DataTypes.STRING,
       description: DataTypes.STRING,
-      price: DataTypes.INTEGER,
-      hours: DataTypes.INTEGER,
+      price: DataTypes.NUMERIC,
+      hours: DataTypes.NUMERIC,
     },
     {}
   );
